% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rforecastio}
\alias{plot.rforecastio}
\title{Plot method for rforecastio objects:}
\usage{
\method{plot}{rforecastio}(object, ..., readings = c("hourly", "minutely",
  "daily"))
}
\arguments{
\item{object}{an rforecastio object}

\item{readings}{specify which readings to plot. will plot all available by default}

\item{\dots}{ignored}
}
\value{
frame grob
}
\description{
Uses ggplot2 & grid.arrange to produce graphs for rforecastio objects
}

