% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fio-print.r
\name{print.rforecastio}
\alias{print.rforecastio}
\title{Slightly more human-readable output}
\usage{
\method{print}{rforecastio}(x, ...)
}
\arguments{
\item{x}{an rforecastio object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output
}

