% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastio-package.r
\docType{package}
\name{Rforecastio}
\alias{Rforecastio}
\alias{Rforecastio-package}
\title{Tools to Work with the Dark Sky API}
\description{
Tools to Work with the Dark Sky API
}
\author{
Bob Rudis (bob@rud.is)
}

