% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rforecastio.R
\name{forecastio_api_key}
\alias{forecastio_api_key}
\title{Get or set FORECASTIO_API_KEY value}
\usage{
forecastio_api_key(force = FALSE, time.formatter = as.POSIXct, ...)
}
\arguments{
\item{force}{force setting a new ForecastIO API key for the current environment?}
}
\value{
atomic character vector containing the ForecastIO API key
}
\description{
The API wrapper functions in this package all rely on a ForecastIO API
key residing in the environment variable \code{FORECASTIO_API_KEY}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}

