% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fio-plot.r
\name{plot.rforecastio}
\alias{plot.rforecastio}
\title{Plot method for rforecastio xs:}
\usage{
\method{plot}{rforecastio}(x, ..., readings = c("hourly", "minutely",
  "daily"))
}
\arguments{
\item{x}{an rforecastio x}

\item{readings}{specify which readings to plot. will plot all available by default}

\item{\dots}{ignored}
}
\value{
frame grob
}
\description{
Uses ggplot2 & grid.arrange to produce graphs for rforecastio objects
}
\note{
only forecast/readings components of \code{x} that have more than one observation
  will be plotted
}

