% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rforecastio-package.R
\docType{package}
\name{Rforecastio}
\alias{Rforecastio}
\alias{Rforecastio-package}
\title{An R wrapper for the Forecast API}
\description{
An R wrapper for the Forecast API
}
\author{
Bob Rudis (@hrbrmstr)
}

