% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastio-package.r
\docType{package}
\name{Rforecastio}
\alias{Rforecastio}
\alias{Rforecastio-package}
\title{An R wrapper for the Forecast API}
\description{
An R wrapper for the Forecast API
}
\author{
Bob Rudis (@hrbrmstr)
}

